telegram_bot/requirements.txt
aiogram==3.4.0
aiohttp==3.9.1
python-dotenv==1.0.0
httpx==0.26.0